refactor useGridState to completely abstract control state
