+++ packages/grid/_modules_/grid/hooks/features/core/useGridState.ts
@@ -2,7 +2,6 @@ import * as React from 'react';
 import { GRID_STATE_CHANGE } from '../../../constants/eventsConstants';
 import { GridApiRef } from '../../../models/api/gridApiRef';
 import { GridStateChangeParams } from '../../../models/params/gridStateChangeParams';
-import { isDeepEqual } from '../../../utils/utils';
 import { GridState } from './gridState';
 import { useGridApi } from './useGridApi';
 
@@ -20,53 +19,14 @@ export const useGridState = (
     (stateUpdaterFn: (oldState: GridState) => GridState) => {
       const newState = stateUpdaterFn(apiRef.current.state);
       const hasChanged = apiRef.current.state !== newState;
-      let shouldUpdate = true;
-      const updatedStateIds: string[] = [];
-      const controlStateMap = apiRef.current.controlStateRef.current!;
-
-      if (hasChanged && apiRef.current.controlStateRef) {
-        Object.keys(controlStateMap).forEach((stateId) => {
-          const controlState = controlStateMap[stateId];
-          const oldState = controlState.stateSelector(apiRef.current.state);
-          const newSubState = controlState.stateSelector(newState);
-          const hasSubStateChanged = !isDeepEqual(oldState, newSubState);
-
-          if (updatedStateIds.length >= 1 && hasSubStateChanged) {
-            // Each hook modify its own state and it should not leak
-            // Events are here to forward to other hooks and apply changes.
-            // You are trying to update several states in a no isolated way.
-            throw new Error(
-              `You're not allowed to update several sub-state in one transaction. You already updated ${updatedStateIds[0]}, therefore, you're not allowed to update ${controlState.stateId} in the same transaction.`,
-            );
-          }
-
-          if (hasSubStateChanged) {
-            if (controlState.propOnChange) {
-              const newModel = controlState.mapStateToModel
-                ? controlState.mapStateToModel(newSubState)
-                : newSubState;
-
-              if (
-                (!controlState.mapStateToModel && controlState.propModel !== newModel) ||
-                (controlState.mapStateToModel && !isDeepEqual(controlState.propModel, newModel))
-              ) {
-                controlState.propOnChange(newModel);
-                shouldUpdate = controlState.propModel === undefined;
-              }
-            } else if (controlState.propModel !== undefined) {
-              const oldModel = controlState.mapStateToModel
-                ? controlState.mapStateToModel(oldState)
-                : oldState;
-              shouldUpdate = controlState.mapStateToModel
-                ? !isDeepEqual(oldModel, controlState.propModel)
-                : oldModel !== controlState.propModel;
-            }
-            if (shouldUpdate) {
-              updatedStateIds.push(controlState.stateId);
-            }
-          }
-        });
+      if (!hasChanged) {
+        // We always assign it as we mutate rows for perf reason.
+        apiRef.current.state = newState;
+        return false;
       }
+
+      const { shouldUpdate, postUpdate } = apiRef.current.applyControlStateConstraint(newState);
+
       if (!shouldUpdate) {
         return false;
       }
@@ -77,17 +37,7 @@ export const useGridState = (
         const params: GridStateChangeParams = { api: apiRef.current, state: newState };
         apiRef.current.publishEvent(GRID_STATE_CHANGE, params);
 
-        updatedStateIds.forEach((stateId) => {
-          if (controlStateMap[stateId].onChangeCallback) {
-            const model =
-              controlStateMap[stateId].mapStateToModel != null
-                ? controlStateMap[stateId].mapStateToModel!(
-                    controlStateMap[stateId].stateSelector(newState),
-                  )
-                : controlStateMap[stateId].stateSelector(newState);
-            controlStateMap[stateId].onChangeCallback!(model);
-          }
-        });
+        postUpdate();
       }
 
       return hasChanged;
