+++ packages/grid/_modules_/grid/models/api/gridControlStateApi.ts
@@ -1,5 +1,10 @@
+import { GridState } from '../../hooks/features/core/gridState';
 import { ControlStateItem } from '../controlStateItem';
 
 export interface GridControlStateApi {
   registerControlState: <TModel, TState>(controlState: ControlStateItem<TModel, TState>) => void;
+  applyControlStateConstraint: (state: GridState) => {
+    shouldUpdate: boolean;
+    postUpdate: () => void;
+  };
 }
